Move classes to proper package
