+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/InstanceSubmitter.kt
@@ -3,8 +3,8 @@ package org.odk.collect.android.instancemanagement
 import org.odk.collect.analytics.Analytics
 import org.odk.collect.android.analytics.AnalyticsEvents
 import org.odk.collect.android.application.Collect
-import org.odk.collect.android.upload.FormUploadException
-import org.odk.collect.android.upload.InstanceServerUploader
+import org.odk.collect.android.instancemanagement.send.FormUploadException
+import org.odk.collect.android.instancemanagement.send.InstanceServerUploader
 import org.odk.collect.android.utilities.FormsRepositoryProvider
 import org.odk.collect.android.utilities.InstanceAutoDeleteChecker
 import org.odk.collect.android.utilities.InstancesRepositoryProvider
