+++ collect_app/src/main/java/org/odk/collect/android/instancemanagement/send/FormUploadException.kt
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.odk.collect.android.upload
+package org.odk.collect.android.instancemanagement.send
 
 import android.net.Uri
 import java.lang.Exception
