+++ collect_app/src/main/java/org/odk/collect/android/utilities/FormsUploadResultInterpreter.kt
@@ -1,9 +1,8 @@
 package org.odk.collect.android.utilities
 
 import android.content.Context
-import org.odk.collect.android.R
+import org.odk.collect.android.instancemanagement.send.FormUploadException
 import org.odk.collect.android.instancemanagement.userVisibleInstanceName
-import org.odk.collect.android.upload.FormUploadException
 import org.odk.collect.errors.ErrorItem
 import org.odk.collect.forms.instances.Instance
 import org.odk.collect.strings.localization.getLocalizedString
